+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changelog
 1.x (unreleased)
 ~~~~~~~~~~~~~~~~
 
+* Return an appropriate error message when the client is not connected.
+
 * Set the ``PGAPPNAME`` used for the PostgreSQL connection, in local mode.
 
 * Close PostgreSQL connections on exit, in local mode.